requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
altair==5.4.1
anaconda==0.0.1.1
arrow==1.3.0
astroid==3.3.3
asttokens==2.4.1
asyncssh==2.17.0
atomicwrites==1.4.1
attrs==24.2.0
autopep8==2.0.4
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
blinker==1.8.2
boto3==1.35.24
botocore==1.35.24
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.8
docstring-to-markdown==0.15
docutils==0.21.2
easyocr==1.7.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
flake8==7.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
huggingface-hub==0.25.0
idna==3.7
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
intel-openmp==2021.4.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.4.1
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.6.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
numpydoc==1.8.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PV-Hawk-Viewer==1.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
pycodestyle==2.12.1
pycparser==2.22
pydeck==0.9.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.4.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.0
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyproj==3.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
pyuca==1.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
roboflow==1.1.45
rope==1.13.0
rpds-py==0.20.0
Rtree==1.3.0
s3transfer==0.10.2
safetensors==0.4.5
sahi==0.11.18
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Shapely==1.8.5.post1
shiboken6==6.7.2
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.0
spyder-kernels==3.0.0
stack-data==0.6.3
streamlit==1.38.0
superqt==0.6.7
sympy==1.13.2
tabulate==0.9.0
tbb==2021.13.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
text-unidecode==1.3
textdistance==4.6.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
three-merge==0.1.1
tifffile==2024.9.20
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.86
ultralytics-thop==2.0.6
urllib3==2.2.2
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
whatthepatch==1.0.6
wrapt==1.16.0
yapf==0.40.2
yarl==1.11.1
yolov5==7.0.13
zipp==3.20.2
